% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqID_maker_addin.R
\name{cheatsheet}
\alias{cheatsheet}
\title{baRcodeR Cheatsheet}
\usage{
cheatsheet()
}
\description{
This addin links to a downloadable PDF version of the baRcodeR cheatsheet.
}
\examples{
if(interactive()){
library(baRcodeR)
cheatsheet()
}
}
