% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelHierarchy.R
\name{label_hier_maker}
\alias{label_hier_maker}
\title{Make hierarchical barcode labels}
\usage{
label_hier_maker(user = F, hierarchy, end = NULL, digits = 2)
}
\arguments{
\item{user}{logical. Should the function be used interactively? Default is
false. When true, all other arguments can be empty.}

\item{hierarchy}{list. A list with each element consisting of three members
a vector of three elements (string, beginning value, end value). See examples.
Used only when \code{user=F})}

\item{end}{character. A string to be appended to end of each label.}

\item{digits}{numerical. Default is 2. Number of digits to be printed. This
will apply to all levels when \code{user=F}. When the numeric value of the
label has a greater number of digits than \code{digits}, \code{digits} is
automatically increased for the entire level.}
}
\value{
data frame of text labels in the first column, and level labels in
separate columns.
}
\description{
Makes hierarchical barcode labels from interactive user input or from list
 with string input, beginning and ending numerical values.
}
\examples{

\dontrun{
## for interactive mode
label_hier_maker(user = T)
}

## how to make hierarchy list

## create vectors for each level in the order string_prefix, beginning_value,
## end_value and combine in list

a <- c("a", 3, 6)
b <- c("b", 1, 3)
c <- list(a, b)
Labels <- label_hier_maker(hierarchy = c)
Labels

## add string at end of each label
Labels <- label_hier_maker(hierarchy = c, end = "end")
Labels

}
