% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelMaker.R
\name{label_maker}
\alias{label_maker}
\title{Makes list for barcode labels in environment}
\usage{
label_maker(user = F, string = NULL, level, digits = 3)
}
\arguments{
\item{user}{logical. Default is FALSE. If the function is run with user
prompts or with variables already specified}

\item{string}{character. Text string for barcode. Default null.}

\item{level}{integer string. Defines the numerical values to be appended
to the character string. Can be any sequence of numbers, not just sequential}

\item{digits}{integer. The number of digits to print for the numbers
appended to the text string. Default is 3.}
}
\value{
a string of characters
}
\description{
Takes user prompts or variables to create  text labels consisting of a text
string and numbers (string001, string002 ... ...) . This function will only
create simple text labels compared to \code{\link{label_hierarchy}} that
creates a hierarchical list of labels. The created output will then be fed
into \code{\link{create_PDF}} to create 2-D labels in PDF.
}
\details{
When the function is used with user prompts, the levels are the sequence
between the starting and ending number. When \code{user=F}, it is possible
to set custom sequences of numbers for the levels. See example.
}
\examples{


## sequential string of numbers in label
Labels<-label_maker(string="string", level = c(1:5), digits=2)
Labels
## can also use nonsequential strings in input for levels
level<-c(1:5, 8:10, 999:1000)
Labels<-label_maker(string="string", level = level, digits=4)
Labels

\dontrun{
## function using user prompt does not use any of the other parameters
Labels<-label_maker(user=T)
Labels
}
}
\seealso{
\code{\link{label_hierarchy}}
}
