% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqID_maker_addin.R
\name{make_labels}
\alias{make_labels}
\title{baRcodeR GUI}
\usage{
make_labels()
}
\description{
This addin will allow you to interactive create ID codes and generate PDF files
of QR codes.
}
\examples{
if(interactive()){
library(baRcodeR)
make_labels()
}
}
