% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelHierarchy.R
\name{uniqID_hier_maker}
\alias{uniqID_hier_maker}
\title{Make hierarchical ID codes}
\usage{
uniqID_hier_maker(user = F, hierarchy, end = NULL, digits = 2)
}
\arguments{
\item{user}{logical. Run function using interactive mode (prompts user for 
parameter values). Default is \code{FALSE}}

\item{hierarchy}{list. A list with each element consisting of three members
a vector of three elements (string, beginning value, end value). See examples.
Used only when \code{user=F})}

\item{end}{character. A string to be appended to end of each label.}

\item{digits}{numerical. Default is \code{2}. Number of digits to be printed, 
adding leading 0s as needed. This will apply to all levels when \code{user=F}. 
When the numeric value of the label has a greater number of digits than 
\code{digits}, \code{digits} is automatically increased for the entire level.}
}
\value{
data.frame of text labels in the first column, with additional columns 
for each level defined by the user.
}
\description{
Generate hierarchical ID codes for barcode labels. Use \code{\link{uniqID_maker}} 
for sequential single-level labels. Can be run interactively 
prompting user for input. The data.frame output can be used as input 
for \code{\link{create_PDF}} to create printable barcode labels.
}
\examples{
\dontrun{
## for interactive mode
uniqID_hier_maker(user = T)
}

## how to make hierarchy list

## create vectors for each level in the order string_prefix, beginning_value,
## end_value and combine in list

a <- c("a", 3, 6)
b <- c("b", 1, 3)
c <- list(a, b)
Labels <- uniqID_hier_maker(hierarchy = c)
Labels

## add string at end of each label
Labels <- uniqID_hier_maker(hierarchy = c, end = "end")
Labels

}
\seealso{
\code{\link{uniqID_maker}}
}
