% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_createPDF.R
\name{custom_create_PDF}
\alias{custom_create_PDF}
\title{Make qr codes and print to stickers}
\usage{
custom_create_PDF(user = F, Labels = NULL, name = "LabelsOut",
  ErrCorr = "H", Fsz = 2.5, Across = T, ERows = 0, ECols = 0,
  trunc = T, numrow = 20, numcol = 4, height_margin = 0.5,
  width_margin = 0.25, cust_spacing = F, x_space = 215)
}
\arguments{
\item{user}{logical. Whether to run function as interactive. Default is \code{FALSE}}

\item{Labels}{vector or data frame.}

\item{name}{character. Name of pdf output file. Default is "LabelsOut"}

\item{ErrCorr}{the error correction value. Level of damage from low to high:
L, M, Q, H. Default is \code{"H"}}

\item{Fsz}{numerical. Set font size. A number between 2.2 and 4.7. Depending
on the length of the label, there may not be enough space to print the entire
 label using bigger font sizes. Default font size is 2.5}

\item{Across}{logical. When true, print labels in rows. When false, print
labels in columns. Default is \code{TRUE}.}

\item{ERows}{number of rows to skip. Default is 0.}

\item{ECols}{number of columns to skip. Default is 0.}

\item{trunc}{logical. Text is split into rows to prevent cutoff when labels
are long. Default is \code{TRUE}.}

\item{numrow}{numerical. Number of rows per page. Default is 20 rows per page.}

\item{numcol}{numerical. Number of columns per page. Default is 4 columns
per page.}

\item{height_margin}{numerical. The height margin of the page (in inches).
Default is 0.5 inches.}

\item{width_margin}{numerical. The width margin of the page (in inches).
Default is 0.25 inches.}

\item{cust_spacing}{logical. Default is \code{FALSE}. If spacing between qrcode and
label should be changed.}

\item{x_space}{numerical. An integer between 190 - 250. This sets the distance
between the qrcode and the label. Default is 215. This parameter is only
used when \code{cust_spacing = T}.}
}
\value{
pdf file that is saved to the working directory containing QR codes.
}
\description{
This function will take in a data frame of labels and produce a pdf of QR
codes which can then be printed. The pdf setup is for the ULINE 1.75X1/2
WEATHER RESISTANT LABEL for laser printer; Item # S-19297 (uline.ca)
}
\seealso{
\code{\link{create_PDF}}
}
