% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{theme_empty}
\alias{theme_empty}
\title{Internal functions}
\usage{
theme_empty(base_size = 12, base_family = "")
}
\description{
Function theme_empty is a code snippet to use in \code{\link{creatPDF}} to remove all lines from plots when creating the barcode PDF file. This function is not exported.
}
